# docs/api/gateway/client.openapi.yaml
...
openapi: 3.0.3
info:
    title: ClientGatewayService API
    description: |-
        ClientGatewayService describes the authenticated REST API surface that the
         web frontend talks to via api-client at /api/v2/*.

         Auth is handled by gateway middleware — the user_id is injected from the
         Firebase token, NEVER sent by the client. Request messages here reflect
         what the frontend actually sends (no user_id fields).
    version: 0.0.1
paths:
...
    /registry:
        get:
            tags:
                - ClientGatewayService
            description: ===================== Registry (Unauthenticated, but on api-client) =====================
            operationId: ClientGatewayService_GetPluginRegistry
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PluginRegistryResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /registry/categories:
        get:
            tags:
                - ClientGatewayService
            operationId: ClientGatewayService_ListCategories
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListCategoriesGatewayResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /registry/plugins:
        get:
            tags:
                - ClientGatewayService
            operationId: ClientGatewayService_GetPluginRegistryPlugins
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PluginRegistryResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /registry/plugins/{id}:
        get:
            tags:
                - ClientGatewayService
            operationId: ClientGatewayService_GetPlugin
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PluginManifest'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
    /registry/sources:
        get:
            tags:
                - ClientGatewayService
            operationId: ClientGatewayService_ListSources
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListSourcesGatewayResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
...
        ConfigFieldDependency:
            type: object
            properties:
                fieldKey:
                    type: string
                values:
                    type: array
                    items:
                        type: string
        ConfigFieldOption:
            type: object
            properties:
                value:
                    type: string
                label:
                    type: string
        ConfigFieldSchema:
            type: object
            properties:
                key:
                    type: string
                label:
                    type: string
                description:
                    type: string
                fieldType:
                    enum:
                        - CONFIG_FIELD_TYPE_UNSPECIFIED
                        - CONFIG_FIELD_TYPE_STRING
                        - CONFIG_FIELD_TYPE_NUMBER
                        - CONFIG_FIELD_TYPE_BOOLEAN
                        - CONFIG_FIELD_TYPE_SELECT
                        - CONFIG_FIELD_TYPE_MULTI_SELECT
                        - CONFIG_FIELD_TYPE_KEY_VALUE_MAP
                        - CONFIG_FIELD_TYPE_DYNAMIC_SELECT
                    type: string
                    format: enum
                required:
                    type: boolean
                defaultValue:
                    type: string
                options:
                    type: array
                    items:
                        $ref: '#/components/schemas/ConfigFieldOption'
                validation:
                    $ref: '#/components/schemas/ConfigFieldValidation'
                dependsOn:
                    $ref: '#/components/schemas/ConfigFieldDependency'
                keyOptions:
                    type: array
                    items:
                        $ref: '#/components/schemas/ConfigFieldOption'
                valueOptions:
                    type: array
                    items:
                        $ref: '#/components/schemas/ConfigFieldOption'
                dynamicSource:
                    type: string
                valueDynamicSource:
                    type: string
        ConfigFieldValidation:
            type: object
            properties:
                minLength:
                    type: integer
                    format: int32
                maxLength:
                    type: integer
                    format: int32
                pattern:
                    type: string
                minValue:
                    type: number
                    format: double
                maxValue:
                    type: number
                    format: double
...
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
        IntegrationAction:
            type: object
            properties:
                id:
                    type: string
                label:
                    type: string
                description:
                    type: string
                icon:
                    type: string
        IntegrationManifest:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                description:
                    type: string
                icon:
                    type: string
                authType:
                    enum:
                        - INTEGRATION_AUTH_TYPE_UNSPECIFIED
                        - INTEGRATION_AUTH_TYPE_OAUTH
                        - INTEGRATION_AUTH_TYPE_API_KEY
                        - INTEGRATION_AUTH_TYPE_APP_SYNC
                        - INTEGRATION_AUTH_TYPE_PUBLIC_ID
                    type: string
                    format: enum
                enabled:
                    type: boolean
                docsUrl:
                    type: string
                setupTitle:
                    type: string
                setupInstructions:
                    type: string
                apiKeyLabel:
                    type: string
                apiKeyHelpUrl:
                    type: string
                marketingDescription:
                    type: string
                features:
                    type: array
                    items:
                        type: string
                iconType:
                    type: string
                iconPath:
                    type: string
                isTemporarilyUnavailable:
                    type: boolean
                actions:
                    type: array
                    items:
                        $ref: '#/components/schemas/IntegrationAction'
...
        ListCategoriesGatewayResponse:
            type: object
            properties:
                categories:
                    type: array
                    items:
                        type: string
...
        ListSourcesGatewayResponse:
            type: object
            properties:
                sources:
                    type: array
                    items:
                        $ref: '#/components/schemas/PluginManifest'
...
        PluginManifest:
            type: object
            properties:
                id:
                    type: string
                type:
                    enum:
                        - PLUGIN_TYPE_UNSPECIFIED
                        - PLUGIN_TYPE_SOURCE
                        - PLUGIN_TYPE_ENRICHER
                        - PLUGIN_TYPE_DESTINATION
                    type: string
                    format: enum
                name:
                    type: string
                description:
                    type: string
                icon:
                    type: string
                configSchema:
                    type: array
                    items:
                        $ref: '#/components/schemas/ConfigFieldSchema'
                requiredIntegrations:
                    type: array
                    items:
                        type: string
                enabled:
                    type: boolean
                enricherProviderType:
                    type: integer
                    format: int32
                destinationType:
                    type: integer
                    format: int32
                marketingDescription:
                    type: string
                features:
                    type: array
                    items:
                        type: string
                transformations:
                    type: array
                    items:
                        $ref: '#/components/schemas/Transformation'
                useCases:
                    type: array
                    items:
                        type: string
                externalUrlTemplate:
                    type: string
                requiredTier:
                    type: string
                category:
                    type: string
                sortOrder:
                    type: integer
                    format: int32
                isPremium:
                    type: boolean
                popularityScore:
                    type: integer
                    format: int32
                iconType:
                    type: string
                iconPath:
                    type: string
                isTemporarilyUnavailable:
                    type: boolean
                allowMultipleInstances:
                    type: boolean
                stage:
                    enum:
                        - PIPELINE_STAGE_UNSPECIFIED
                        - PIPELINE_STAGE_GATE
                        - PIPELINE_STAGE_ENRICHMENT
                        - PIPELINE_STAGE_METRICS
                        - PIPELINE_STAGE_CONTEXT_AI
                        - PIPELINE_STAGE_INPUT
                        - PIPELINE_STAGE_VIZ
                    type: string
                    description: |-
                        PipelineStage groups boosters in the pipeline editor UI.
                         Only meaningful for type == PLUGIN_TYPE_ENRICHER.
                    format: enum
                supportsNonBlocking:
                    type: boolean
                    description: |-
                        supportsNonBlocking indicates this enricher can run in non-blocking mode:
                         the pipeline continues to destinations while the enricher waits for user
                         input, then destinations are updated (not duplicated) when input is resolved.
        PluginRegistryResponse:
            type: object
            properties:
                sources:
                    type: array
                    items:
                        $ref: '#/components/schemas/PluginManifest'
                enrichers:
                    type: array
                    items:
                        $ref: '#/components/schemas/PluginManifest'
                destinations:
                    type: array
                    items:
                        $ref: '#/components/schemas/PluginManifest'
                integrations:
                    type: array
                    items:
                        $ref: '#/components/schemas/IntegrationManifest'
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...
        Transformation:
            type: object
            properties:
                field:
                    type: string
                label:
                    type: string
                before:
                    type: string
                after:
                    type: string
                visualType:
                    type: string
                afterHtml:
                    type: string
...